# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "A RESTful API built with Go, Gin, GORM, and PostgreSQL using Clean Architecture principles",
        "title": "Go Clean Architecture API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/example": {
            "post": {
                "description": "Create a new example with a description",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "examples"
                ],
                "summary": "Create a new example",
                "parameters": [
                    {
                        "description": "Example object with description",
                        "name": "example",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.Example"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successfully created example",
                        "schema": {
                            "$ref": "#/definitions/controller.ExampleResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "$ref": "#/definitions/controller.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to save example",
                        "schema": {
                            "$ref": "#/definitions/controller.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/example/{exampleId}": {
            "get": {
                "description": "Retrieve a specific example by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "examples"
                ],
                "summary": "Get an example by ID",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"1\"",
                        "description": "Example ID",
                        "name": "exampleId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved example",
                        "schema": {
                            "$ref": "#/definitions/controller.ExampleResponse"
                        }
                    },
                    "404": {
                        "description": "Example not found",
                        "schema": {
                            "$ref": "#/definitions/controller.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "controller.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Error message"
                }
            }
        },
        "controller.ExampleResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/entity.Example"
                }
            }
        },
        "entity.Example": {
            "type": "object",
            "required": [
                "description"
            ],
            "properties": {
                "description": {
                    "type": "string",
                    "example": "This is an example"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                }
            }
        }
    }
}